move out of bullpen
