+++ bullpen/src/bullpen/__init__.py
@@ -1,5 +1,5 @@
-from . import api, logging, timing, util
+from . import api, logging, util
 
 PLUGIN_GROUP = "bullpen.mlbled.plugin"
 
-__all__ = ["api", "logging", "timing", "util"]
+__all__ = ["api", "logging", "util"]
